% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGR.R
\name{BGR}
\alias{BGR}
\alias{BGR.EE}
\alias{BGR.EI}
\alias{BGR.IE}
\title{Bivariate gamma regression}
\usage{
BGR(modelName = c("EE", "EI", "IE"), response, data, l1, l2, l3, l4,
  expo = NULL, maxit = 100, tol = 1e-05, Aitken = TRUE,
  verbose = TRUE)

BGR.EE(data, response, l1, l2, l3, l4, expo = NULL, maxit = 200,
  tol = 1e-04, Aitken = FALSE, verbose = TRUE)

BGR.EI(data, response, l1, l2, l3, expo = NULL, maxit = 200,
  tol = 1e-04, Aitken = FALSE, verbose = TRUE)

BGR.IE(data, response, l4, expo = NULL, maxit = 200, tol = 1e-04,
  Aitken = FALSE, verbose = TRUE)
}
\arguments{
\item{modelName}{A character string indicating which model to be fitted. Need to be one of "EE", "EI", "IE".}

\item{response}{A vector of character strings indicating which variables in the data are treated as response or dependent variables.}

\item{data}{A matrix or data frame of observations. Categorical variables are allowed as covariates.}

\item{l1}{A regression formula for the \eqn{latex}{\alpha_1} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{l2}{A regression formula for the \eqn{latex}{\alpha_2} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{l3}{A regression formula for the \eqn{latex}{\alpha_3} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{l4}{A regression formula for the \eqn{latex}{\beta} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{expo}{Exposure variable}

\item{maxit}{A parameter that controls the number of maximum iteration in the EM algorithm. The default is 100.}

\item{tol}{A parameter that controls the convergence tolerance in the EM algorithm. The default is 1e-5.}

\item{Aitken}{A parameter that contols whether Aitken acceleration is used.}

\item{verbose}{A logical controlling whether estimations in each EM iteration are shown in the fitting process. The default is TRUE.}
}
\value{
An object of class \code{BGR} providing the estimation results.
  The details of the output components are:
  \item{call}{The matched call.}
  \item{coef}{The estimated coefficients.}
  \item{alpha1}{The estimated alpha1 values.}
  \item{alpha2}{The estimated alpha2 values.}
  \item{alpha3}{The estimated alpha3 values.}
  \item{beta}{The estimated beta values.}
  \item{trace}{All estimated coefficients and alpha, beta values in the EM algorithm.}
  \item{fitted}{The fitted values of the regression.}
  \item{loglike}{The final estimated maximum log-likelihood value.}
  \item{llseq}{The sequence of log-likelihood values in the EM algorithm fitting process.}
  \item{param.number}{Number of estimated parameters.}
  \item{AIC}{AIC values.}
  \item{BIC}{BIC values.}
  \item{iter}{Total iteration numbers.}
  \item{formula}{The formulas used in the regression.}
  \item{y}{The input response data.}
  \item{Model.Matrix}{The used model matrix for each regression formula.}
}
\description{
Bivariate gamma regression models for three model types: EE, EI, IE. Models are estimated by EM algorithm.
}
\examples{
m1 <- BGR(modelName = "EE",
          data   = regsim,
          response=c("y1","y2"),
          l1     = ~ w1 + w2,
          l2     = ~ w2 + w3,
          l3     = ~ w1 + w2 + w3,
          l4     = ~ w1 + w2 + w3,
          expo   = NULL,
          maxit  = 1000,
          tol    = 1e-6,
          Aitken = FALSE,
          verbose= TRUE)
}
