% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAweight_bic.R
\name{BMAweight_bic}
\alias{BMAweight_bic}
\title{Approximate model posterior probability using BIC.}
\usage{
BMAweight_bic(BICvec)
}
\arguments{
\item{BICvec}{a vector of all BIC values used to calculate model weight}
}
\value{
the approximated posterior model probability, i.e. model weight used in BMA.
}
\description{
This function calculates approximately model posterior probability given BIC values, which can then be used for BMA.
}
\examples{
BMAweight_bic(c(10,11,9,15,9.5))

}
