% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copreg.gamma.R
\name{copreg.gamma}
\alias{copreg.gamma}
\title{Bivariate Copula Regression with Gamma Marginal Distributions}
\usage{
copreg.gamma(f1, f2, copula, data, optim.method = "BFGS",
  optim.control = list(fnscale = -1, trace = 1))
}
\arguments{
\item{f1}{A regression formula for the first marginal GLM.}

\item{f2}{A regression formula for the second marginal GLM.}

\item{copula}{A copula specified from the \code{copula} package.}

\item{data}{A matrix or data frame of observations. Categorical variables are allowed as covariates.}

\item{optim.method}{Method for the \code{optim} function.}

\item{optim.control}{Parameters that controls the \code{optim} function used for maximum likelihood.}
}
\value{
An object of class \code{copreg} providing the estimation results.
  The details of the output components are:
  \item{call}{The matched call.}
  \item{coefficients}{The estimated coefficients.}
  \item{copula.param}{The estimated copula parameters.}
  \item{copula}{The specified copula.}
  \item{fitted.values}{The fitted values of the regression.}
  \item{residuals}{The residuals from fitted values.}
  \item{loglike}{The final estimated maximum log-likelihood value.}
  \item{df}{Number of estimated parameters.}
  \item{AIC}{AIC values.}
  \item{BIC}{BIC values.}
  \item{formula}{The formulas used in the regression.}
  \item{y}{The input response data.}
  \item{n}{The number of observations in the data.}
  \item{Model.Matrix}{The used model matrix for each regression formula.}
}
\description{
Bivariate copula regression models where the marginal distributions are of Gamma GLMs.
}
\examples{
data("simdat.mcgr")
res <- copreg.gamma(f1 = y1 ~ x1+x2,
                    f2 = y2 ~ x1+x2,
                    copula = copula::gumbelCopula(dim=2),
                    data = simdat.mcgr)

}
