% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbivgamma.R
\name{dbivgamma}
\alias{dbivgamma}
\title{Density of Bivariate Gamma Distribution}
\usage{
dbivgamma(y, alpha, beta, log = FALSE)
}
\arguments{
\item{y}{vector of quantiles}

\item{alpha}{vector of alpha values in bivariate gamma distribution}

\item{beta}{value of beta in bivariate gamma distribution}

\item{log}{logical; if TRUE, probabilities p are given as log(p). Default is FALSE.}
}
\description{
Density of bivariate gamma distribution given parameters values
}
\examples{
dbivgamma(c(1,1), alpha=c(0.3, 0.4, 0.5), beta=1, log = TRUE)

\dontrun{
c1 <- seq(0, 15, by=0.1)[-1]
c2 <- seq(0, 15, by=0.1)[-1]
restable <- matrix(0, nrow = length(c1), ncol=length(c2))
for (i in c(1:length(c1))){
 for (j in c(1:length(c2))){
   restable[i,j] <- dbivgamma(c(c1[i], c2[j]),
                              alpha = c(0.5,0.5,0.5),
                              beta = 1,log=TRUE)
 } }
rownames(restable) <- NULL
colnames(restable) <- NULL
require(lattice)
new.palette=colorRampPalette(c("white","yellow","red"),space="rgb")
levelplot(restable, col.regions=new.palette(20), xlab="Y_1", ylab="Y_2")
}
}
