% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbivpois.R
\name{dbivpois}
\alias{dbivpois}
\title{Density of bivariate Poisson distribution}
\usage{
dbivpois(y, lambda = c(1, 1, 1), log = FALSE, method = "analytical")
}
\arguments{
\item{y}{a vector}

\item{lambda}{vector of lambda values for lambda1, lambda2, lambda3.}

\item{log}{whether the calculated density is on log scale}

\item{method}{"analytical" or "recursive"}
}
\description{
Density of bivariate Poisson distribution
}
\note{
This function is borrowed from "bivpois" R package which is no longer available on CRAN
}
