% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvpois.R
\name{dmvpois}
\alias{dmvpois}
\title{Estimate of the density for a given vector lambda of multivariate Poisson distribution (MVP)}
\usage{
dmvpois(y, lambda, method, log, n.mc = 10000)
}
\arguments{
\item{y}{a vector of MVP sample}

\item{lambda}{underlying Poisson rates}

\item{method}{"analytical" or "recursive"}

\item{log}{whether the calculated value on log scale}

\item{n.mc}{number of Monte Carlo samples if method='MC"}
}
\value{
Estimate of the probability P(x) for a given vector lambda of Poisson rates
}
\description{
Estimate of the density for a given vector lambda of multivariate Poisson distribution (MVP)
}
\note{
This function is borrowed from "mvrpois" package available from GitHub.
}
\examples{
dmvpois(y=c(2,2), lambda=c(1,1,1), method="recursive", log = F)
# analytical method only works for d>=3, i.e. not working for bivariate case

dmvpois(y=c(2,2,2), lambda=c(1,1,1,1,1,1), method="recursive", log = F)
dmvpois(y=c(2,2,2), lambda=c(1,1,1,1,1,1), method="analytical", log = F)
dmvpois(y=c(2,2,2), lambda=c(1,1,1,1,1,1), method="MC", log = F)

}
