% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingsim.R
\docType{data}
\name{gatingsim}
\alias{gatingsim}
\title{Simulated artificial data with covariates in the gating network.}
\format{A data frame with 500 rows and 6 variables:
\describe{
  \item{y1,y2}{The bivariate response variable.}
  \item{w1,w2,w3}{Three covariates.}
  \item{label}{Their true labels.}
}}
\source{
Hu, S., Murphy, T. B. and O'Hagan, A. (2019) Bivariate Gamma Mixture of Experts Models for Joint Claims Modeling.
}
\usage{
gatingsim
}
\description{
An artificial dataset containing the bivariate response y1, y2, three covariates w1, w2, w3, and the true label.
}
\keyword{datasets}
