% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBGR.R
\name{MBGR}
\alias{MBGR}
\title{Mixture of Bivariate Gamma Clustering/Regressions}
\usage{
MBGR(modelName = c("VC", "VI", "VV", "VE", "CV", "IV", "EV", "EC", "CE"),
  y, data, G, f1, f2, f3, f4, gating, initialization = "mclust",
  maxit = 200, tol = 1e-05, verbose = TRUE)
}
\arguments{
\item{modelName}{A character string indicating which model to be fitted. Need to be one of "EE", "EI", "IE".}

\item{y}{A vector of character strings indicating which variables in the data are treated as response or dependent variables.}

\item{data}{A matrix or data frame of observations. Categorical variables are allowed as covariates.}

\item{G}{An interger vector specifying the number of mixture components (clusters).}

\item{f1}{A regression formula for the \eqn{latex}{\alpha_1} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{f2}{A regression formula for the \eqn{latex}{\alpha_2} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{f3}{A regression formula for the \eqn{latex}{\alpha_3} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{f4}{A regression formula for the \eqn{latex}{\beta} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{gating}{Specifies the gating network in the MoE model, can be "C", "E" or a regression formula.}

\item{initialization}{Specifies initialization method for EM algorithm. The default is "\code{mclust}".}

\item{maxit}{A parameter that controls the number of maximum iteration in the EM algorithm. The default is 100.}

\item{tol}{A parameter that controls the convergence tolerance in the EM algorithm. The default is 1e-5.}

\item{verbose}{A logical controlling whether estimations in each EM iteration are shown in the fitting process. The default is TRUE.}
}
\value{
An object of class \code{BGR} providing the estimation results.
  The details of the output components are:
  \item{call}{The matched call.}
  \item{expert.coef}{The estimated coefficients in the expert network, if exists.}
  \item{gating.coef}{The estimated coefficients in the gating network, if exists.}
  \item{alpha1}{The estimated alpha1 values.}
  \item{alpha2}{The estimated alpha2 values.}
  \item{alpha3}{The estimated alpha3 values.}
  \item{beta}{The estimated beta values.}
  \item{pro}{A vector whose g-th component is the mixing proportion for the g-th component of the mixture model.}
  \item{z}{A matrix whose [i,g]-th entry is the probability that observation i in the data belongs to the g-th group.}
  \item{class}{The classification corresponding to z.}
  \item{fitted.values}{The fitted values of the regression.}
  \item{residuals}{The residuals of the regression}
  \item{loglike}{The final estimated maximum log-likelihood value.}
  \item{ll}{The sequence of log-likelihood values in the EM algorithm fitting process.}
  \item{df}{The number of estimated parameters.}
  \item{AIC}{AIC values.}
  \item{BIC}{BIC values.}
  \item{Hessian}{The Hessian matrix at the estimated values}
  \item{iter}{Total iteration numbers.}
  \item{formula}{The formulas used in the regression.}
  \item{y}{The input response data.}
  \item{n}{The number of observations in the data.}
  \item{gating.model}{The binomial/multinomial regression model in the gating network.}
  \item{Model.Matrix}{The used model matrix for each regression formula.}
}
\description{
Mixture of bivariate gamma regressions, or model-based clustering with bivariage gamma distributions and covariates, for various parsimonious model types. Models are estimated by EM algorithm.
}
\examples{
\donttest{
m1 <- MBGR(modelName = "VV",
           y=c("y1","y2"), data = fullsim, G=2,
           f1     = ~ w1 + w2,
           f2     = ~ w2 + w3,
           f3     = ~ w1 + w2 + w3,
           f4     = ~ w1 + w2 + w3,
           gating = "C",
           verbose = FALSE)
}

}
