% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBPC.R
\name{MBPC}
\alias{MBPC}
\title{Model-based bivariate Poisson clustering}
\usage{
MBPC(G, data, start, maxit = 300, tol = 1e-05, Aitken = FALSE,
  verbose = TRUE)
}
\arguments{
\item{G}{number of mixtures}

\item{maxit}{maximum iteration. Default is 300.}

\item{tol}{tolerance. Default is 1e-5.}

\item{Aitken}{if TRUE, using Aitken acceleration estimate criterion. Defaul is TRUE.}

\item{verbose}{if TRUE, print the iteration results. Default is TRUE.}
}
\description{
Model-based bivariate Poisson clustering
}
