% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBPR.R
\name{MBPR}
\alias{MBPR}
\title{Mixture of bivariate Poisson regression (mixture of experts)}
\usage{
MBPR(G, data, l1, l2, l3, lp = NULL, w = NULL, maxit = 300,
  tol = 1e-06, Aitken = FALSE, verbose = TRUE)
}
\arguments{
\item{l1}{formula for lambda1}

\item{l2}{formula for lambda2}

\item{l3}{formula for lambda3. If NULL, only double Poisson regression. Default is NULL.}

\item{lp}{formula for mixing proportion. If NULL, mixing proportion is not regressed against covariates. Default is NULL.}

\item{w}{weights in GLM (if applicable) such as "exposure". Default is NULL.}

\item{maxit}{maximum iteration. Default is 300.}

\item{tol}{tolerance. Default is 1e-5.}

\item{Aitken}{if TRUE, using Aitken acceleration estimate criterion. Defaul is TRUE.}

\item{verbose}{if TRUE, print the iteration results. Default is TRUE.}

\item{K}{number of mixtures}
}
\description{
Mixture of bivariate Poisson regression (mixture of experts)
}
