% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCGR.R
\name{MCGR}
\alias{MCGR}
\title{Mixture of Copula Regressions with Gamma Marginal Distributions}
\usage{
MCGR(copula, f1, f2, G, data, gating, initialization = "mclust",
  maxit = 300, tol = 1e-05, verbose = FALSE)
}
\arguments{
\item{copula}{A specified copula, from the \code{copula} package. Must be a list of copulas or a list of one copula, matching G}

\item{f1}{A regression formula for the first marginal GLM.}

\item{f2}{A regression formula for the second marginal GLM.}

\item{G}{An interger vector specifying the number of mixture components (clusters).}

\item{data}{A matrix or data frame of observations. Categorical variables are allowed as covariates.}

\item{gating}{Specifies the gating network, can be "C", "E" or a regression formula.}

\item{initialization}{Specifies initialization method for EM algorithm. The default is "\code{mclust}".}

\item{maxit}{A parameter that controls the number of maximum iteration in the EM algorithm. The default is 100.}

\item{tol}{A parameter that controls the convergence tolerance in the EM algorithm. The default is 1e-5.}

\item{verbose}{A logical controlling whether estimations in each EM iteration are shown in the fitting process. The default is TRUE.}
}
\value{
An object of class \code{MCGR} providing the estimation results.
  The details of the output components are:
  \item{call}{The matched call.}
  \item{coefficients}{The estimated coefficients in the expert and gating networks, if exists.}
  \item{copula.param}{The estimated copula parameters.}
  \item{copula}{The specified copula.}
  \item{theta}{All estimated parameter values, in a vector format.}
  \item{pro}{A vector whose g-th component is the mixing proportion for the g-th component of the mixture model.}
  \item{z}{A matrix whose [i,g]-th entry is the probability that observation i in the data belongs to the g-th group.}
  \item{class}{The classification corresponding to z.}
  \item{fitted.values}{The fitted values of the regression.}
  \item{residuals}{The residuals of the regression}
  \item{loglike}{The final estimated maximum log-likelihood value.}
  \item{ll}{The sequence of log-likelihood values in the EM algorithm fitting process.}
  \item{df}{The number of estimated parameters.}
  \item{AIC}{AIC values.}
  \item{BIC}{BIC values.}
  \item{iter}{Total iteration numbers.}
  \item{formula}{The formulas used in the regression.}
  \item{y}{The input response data.}
  \item{n}{The number of observations in the data.}
  \item{gating.model}{The binomial/multinomial regression model in the gating network.}
  \item{Model.Matrix}{The used model matrix for each regression formula.}
}
\description{
Mixture of copula regressions, or model-based clustering with copulas and covariates. Models are estimated by EM algorithm.
}
\examples{
\donttest{
data("simdat.mcgr")
res <- MCGR(copula = list(copula::gumbelCopula(dim=2),
                          copula::frankCopula(dim=2)),
            f1 = y1 ~ x1+x2,
            f2 = y2 ~ x1+x2,
            G = 2,
            gating = "C",
            data = simdat.mcgr,
            verbose = FALSE)
}

}
