% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPR_need to update.R
\name{MMPR}
\alias{MMPR}
\title{Mixture of multivariate Poisson regression (mixture of experts)}
\usage{
MMPR(K = 2, l, lc = NULL, lp = NULL, w = NULL, data, maxit = 300,
  tol = 1e-05, Aitken = TRUE, verbose = TRUE)
}
\arguments{
\item{K}{number of mixtures}

\item{l}{formula for main effect lambda}

\item{lc}{formula for covariance term lambda}

\item{lp}{formula for mixing proportion. If NULL, mixing proportion is not regressed against covariates. Default is NULL.}

\item{w}{weights in GLM (if applicable) such as "exposure". Default is NULL.}

\item{maxit}{maximum iteration. Default is 300.}

\item{tol}{tolerance. Default is 1e-5.}

\item{Aitken}{if TRUE, using Aitken acceleration estimate criterion. Defaul is TRUE.}

\item{verbose}{if TRUE, print the iteration results. Default is TRUE.}
}
\description{
Mixture of multivariate Poisson regression (mixture of experts)
}
\examples{
cor(cbind(health$doctorco, health$prescrib, health$hospadmi))
l <- list(doctorco~sex+age+income,
          prescrib~sex+age+income,
          hospadmi~sex+age+income)
lc <- list(x12~sex+age+income,
           x13~sex+age+income,
           x23~sex+age+income)
lp <- ~sex+age+income
data <- health
}
