% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MBGR}
\alias{plot.MBGR}
\alias{plot.BGR}
\alias{plot.MBGC}
\title{Plot Mixture of Bivariate Gamma Regressions and Clustering Results}
\usage{
\method{plot}{MBGR}(x, what = "classification", col = NULL,
  pch = NULL, xlab = NULL, ylab = NULL, main = FALSE, ...)

\method{plot}{BGR}(x, col = NULL, pch = NULL, xlab = NULL,
  ylab = NULL, main = FALSE, ...)

\method{plot}{MBGC}(x, what = "classification", col = NULL,
  pch = NULL, xlab = NULL, ylab = NULL, main = FALSE, ...)
}
\arguments{
\item{x}{Output from MBGR functions.}

\item{what}{The type of graph requested, either \code{"classification"} or \code{"uncertainty"}.}

\item{col}{The colors of points in clusters.}

\item{pch}{The plotting characters or symbols in clusters.}

\item{xlab, ylab}{Optional labels for the x-axis and the y-axis.}

\item{main}{A logical or \code{NULL} indicating whether or not to add a title to the plot identifying the type of plot.}

\item{...}{Other graphical parameters.}
}
\value{
For BGR model, fitted values are plotted.

  For MBGR model:
  \item{what=="classification"}{a plot showing the clustering labels.}
  \item{what=="uncertainty"}{a plot of classification uncertainty.}
  \item{what=="fitted"}{a plot of fitted values.}
  For MBGR model:
  \item{what=="classification"}{a plot showing the clustering labels.}
  \item{what=="uncertainty"}{a plot of classification uncertainty.}
}
\description{
Plot mixture of bivariate gamma regressions, or model-based clustering with bivariage gamma distributions and covariates: classification, uncertainty
}
\examples{

\donttest{
m1 <- MBGR(modelName = "CE", y=c("y1","y2"),
           data   = fullsim, G=2,
           f1     = ~ w1 + w2,
           f2     = ~ w2 + w3,
           f3     = ~ w1 + w2 + w3,
           f4     = ~ w1 + w2 + w3,
           gating = "C", verbose = FALSE)
plot.MBGR(m1, what="classification")
plot.MBGR(m1, what="uncertainty")
plot.MBGR(m1, what="fitted")
m2 <- BGR(modelName = "EI",
          y = c("y1","y2"), data = fullsim,
          f1     = ~ w1 + w2,
          f2     = ~ w2 + w3,
          f3     = ~ w1 + w2 + w3, verbose = FALSE)
plot.BGR(m2)
}

}
