% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.BGR.R
\name{predict.BGR}
\alias{predict.BGR}
\title{Predict Method for Bivariate Gamma Regressions}
\usage{
\method{predict}{BGR}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted model object of class \code{"BGR"} or \code{"MBGR"}}

\item{newdata}{a data frame that prediction is based upon}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list with components (depends on model class):
  \item{fit}{predictions.}
  \item{alpha1.fit}{predicted alpha1 values.}
  \item{alpha2.fit}{predicted alpha2 values.}
  \item{alpha3.fit}{predicted alpha3 values.}
  \item{beta.fit}{predicted beta values.}
  \item{tau.fit}{predicted mixing proportion values, if covariates enter the gating network.}
}
\description{
Obtains predictions from the fitted bivariate gamma regression models.
}
\examples{

\donttest{
mod1 <- BGR(modelName = "EE",
            y = c("y1","y2"), data = fullsim,
            f1 = ~ w1 + w2,
            f2 = ~ w2 + w3,
            f3 = ~ w1 + w2 + w3,
            f4 = ~ w1 + w2 + w3,
            verbose= FALSE)
fitted1 <- predict(mod1, newdata=fullsim)
plot(fitted1$fit)

mod2 <- BGR(modelName = "EI",
            y = c("y1","y2"), data = fullsim,
            f1     = ~ w1 + w2,
            f2     = ~ w2 + w3,
            f3     = ~ w1 + w2 + w3,
            verbose= FALSE)
fitted2 <- predict(mod2, newdata=fullsim)
mod3 <- BGR(modelName = "IE",
            y = c("y1","y2"), data = fullsim,
            f4     = ~ w1 + w2 + w3,
            verbose= FALSE)
fitted3 <- predict(mod3, newdata=fullsim)
plot(fitted2$fit)
plot(fitted3$fit)
}

}
