% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbivgamma.R
\name{rbivgamma}
\alias{rbivgamma}
\title{Random generation for the bivariate gamma distribution}
\usage{
rbivgamma(n, alpha, beta)
}
\arguments{
\item{n}{number of observations.}

\item{alpha}{the shape parameters of the bivariate gamma distribution. Must be a vector of length 3 represent alpha 1, alpha 2 and alpha 3 respectively.}

\item{beta}{the rate parameter of the bivariate gamma distribution.}
}
\value{
random deviates of length n
}
\description{
This function allows you to generate random bivariate gamma variables given the distribution parameters
}
\examples{
plot(rbivgamma(2000, alpha=c(1,1,1), beta=1))
plot(rbivgamma(2000, c(0.5, 0.5, 0.2),beta=0.001))

}
