% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.mcgr.R
\docType{data}
\name{simdat.mcgr}
\alias{simdat.mcgr}
\title{Simulated artificial data with covariates for `MCGR`.}
\format{A data frame with 1000 rows and 5 variables:
\describe{
  \item{y1,y2}{The bivariate response variable.}
  \item{x1,x2}{Two covariates.}
  \item{label}{Their true labels.}
}}
\source{
Hu, S. and O'Hagan, A. (2019) Copula Averaging for Tail Dependence in General Insurance Claims Data. To appear.
}
\usage{
simdat.mcgr
}
\description{
An artificial dataset containing the bivariate response y1, y2, three covariates x1, x2, and the true label.
}
\keyword{datasets}
