% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BGE.R
\name{summary.BGE}
\alias{summary.BGE}
\alias{print.summary.BGE}
\title{Summarizing Bivariate Gamma Distribution Fits}
\usage{
\method{summary}{BGE}(object, ...)

\method{print}{summary.BGE}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class "\code{BGE}" resulting of a call to \code{BGE}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "\code{summary.BGE}", usually a result of a call to \code{summary.BGE}.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
Summary method for class "\code{BGE}"
}
\examples{
\donttest{
dat <- rbivgamma(1000, alpha = c(1,2,0.5), beta=0.1)
mod <- BGE(data = dat, verbose = FALSE)
summary(mod)
}

}
