% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BGR.R
\name{summary.BGR}
\alias{summary.BGR}
\alias{print.summary.BGR}
\title{Summarizing Bivariate Gamma Regression Model Fits}
\usage{
\method{summary}{BGR}(object, ...)

\method{print}{summary.BGR}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class "\code{BGR}" resulting of a call to \code{BGR}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "\code{summary.BGR}", usually a result of a call to \code{summary.BGR}.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
Summary method for class "\code{BGR}"
}
\examples{
mod <- BGR(modelName = "EE",
           y=c("y1","y2"),
           data   = fullsim,
           f1     = ~ w1 + w2,
           f2     = ~ w2 + w3,
           f3     = ~ w1 + w2 + w3,
           f4     = ~ w1 + w2 + w3,
           verbose= FALSE)
summary(mod)
}
