% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.copreg.gamma.R
\name{summary.copreg.gamma}
\alias{summary.copreg.gamma}
\alias{print.summary.copreg.gamma}
\title{Summarizing Bivariate Copula Regression with Gamma Margins Model Fits}
\usage{
\method{summary}{copreg.gamma}(object, ...)

\method{print}{summary.copreg.gamma}(x, digits = getOption("digits"),
  ...)
}
\arguments{
\item{object}{An object of class "\code{copreg.gamma}" resulting of a call to \code{copreg.gamma}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "\code{summary.copreg.gamma}", usually a result of a call to \code{summary.copreg.gamma}.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
Summary method for class "\code{copreg.gamma}"
}
\examples{

\donttest{
mod <- copreg.gamma(f1 = y1 ~ x1+x2,
                    f2 = y2 ~ x1+x2,
                    copula =  gumbelCopula(dim=2),
                    data = simdat.mcgr)
summary(mod)
}

}
