% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MBGC.R
\name{summary.MBGC}
\alias{summary.MBGC}
\alias{print.summary.MBGC}
\title{Summarizing Mixture of Bivariate Gamma Clustering Model Fits}
\usage{
\method{summary}{MBGC}(object, ...)

\method{print}{summary.MBGC}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class "\code{MBGC}" resulting of a call to \code{MBGC}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "\code{summary.MBGC}", usually a result of a call to \code{summary.MBGC}.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
Summary method for class "\code{MBGC}"
}
\examples{
\donttest{
clust1 <- MBGC(modelName = "CC", y=c("y1","y2"),
               G=2, gating = "C", data=gatingsim)
summary(clust1)
clust2 <- MBGC(modelName = "CI", y=c("y1","y2"),
               G=2, gating = "C", data=gatingsim)
summary(clust2)
clust4 <- MBGC(modelName = "CC", y=c("y1","y2"),
               G=2, gating = ~w1+w2+w3, data=gatingsim)
summary(clust4)
}

}
