% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MBGR.R
\name{summary.MBGR}
\alias{summary.MBGR}
\alias{print.summary.MBGR}
\title{Summarizing Mixture of Bivariate Gamma Regressions Model Fits}
\usage{
\method{summary}{MBGR}(object, ...)

\method{print}{summary.MBGR}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class "\code{MBGR}" resulting of a call to \code{MBGR}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "\code{summary.MBGR}", usually a result of a call to \code{summary.MBGR}.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
Summary method for class "\code{MBGR}"
}
\examples{
mod <- MBGR(modelName = "CE",
            y=c("y1","y2"),
            data   = fullsim,
            G=2,
            f1     = ~ w1 + w2,
            f2     = ~ w2 + w3,
            f3     = ~ w1 + w2 + w3,
            f4     = ~ w1 + w2 + w3,
            gating = "C")
summary(mod)
}
