% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MCGR.R
\name{summary.MCGR}
\alias{summary.MCGR}
\alias{print.summary.MCGR}
\title{Summarizing Mixture of Copula Regressions with Gamma Marginal Distributions Model Fits}
\usage{
\method{summary}{MCGR}(object, ...)

\method{print}{summary.MCGR}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class "\code{MCGR}" resulting of a call to \code{MCGR}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "\code{summary.MCGR}", usually a result of a call to \code{summary.MCGR}.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
Summary method for class "\code{MCGR}"
}
\examples{
\donttest{
mod <- MCGR(copula = list(gumbelCopula(dim=2), frankCopula(dim=2)),
            f1 = y1 ~ x1+x2,
            f2 = y2 ~ x1+x2,
            G = 2,
            gating = "C",
            data=simdat.mcgr)
summary(mod)
}

}
