% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDC.R
\name{TDC}
\alias{TDC}
\title{Tail Dependence Coefficient}
\usage{
TDC(X, lower, method, plot, control)
}
\arguments{
\item{X}{A data matrix}

\item{lower}{Logical; if \code{TRUE} the lower TDC is computed, and upper TDC otherwise.}

\item{method}{Specifies which method to use, integer values from 1 to 4.}

\item{plot}{Logical; if \code{TRUE} the function also plots the corresponding plot during estimation. Only for method 3 and 4.}

\item{control}{A list of control parameters for method 4.}
}
\value{
The estimated tail dependence coefficient value
}
\description{
This function returns the tail dependence coefficients that estimated nonparametrically using empirical copulas.
}
\examples{
data(simdat.mcgr)
TDC(simdat.mcgr[,1:2], method = 1, lower=FALSE)
TDC(simdat.mcgr[,1:2], method = 2, lower=FALSE)
TDC(simdat.mcgr[,1:2], method = 3, lower=FALSE)

}
