% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGE.R
\name{BGE}
\alias{BGE}
\title{Bivariate gamma distribution estimation}
\usage{
BGE(data, maxit = 300, tol = 1e-06, start.value = NULL,
  Aitken = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{A numeric vector, matrix or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.}

\item{maxit}{An integer limits on the number of EM iterations. The default is 300.}

\item{tol}{A value giving relative convergence tolerance for the log-likelihood. The default is 1e-6.}

\item{start.value}{Starting value of the EM algorithm}

\item{Aitken}{logical; indicate whether Aitken Acceleration is used}

\item{verbose}{logical; controls whether summary result of each EM iteration is displayed during the fitting procedure. Default is TRUE.}
}
\value{

}
\description{
This function allows you to estimate bivariate gamma distribution given a data set using EM algorithm
}
