% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGR.R
\name{BGR}
\alias{BGR}
\title{Bivariate Gamma Regression}
\usage{
BGR(modelName = c("EE", "EI", "IE"), y, data, f1, f2, f3, f4,
  maxit = 100, tol = 1e-05, verbose = TRUE)
}
\arguments{
\item{modelName}{A character string indicating which model to be fitted. Need to be one of "EE", "EI", "IE".}

\item{y}{A vector of character strings indicating which variables in the data are treated as response or dependent variables.}

\item{data}{A matrix or data frame of observations. Categorical variables are allowed as covariates.}

\item{f1}{A regression formula for the \eqn{latex}{\alpha_1} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{f2}{A regression formula for the \eqn{latex}{\alpha_2} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{f3}{A regression formula for the \eqn{latex}{\alpha_3} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{f4}{A regression formula for the \eqn{latex}{\beta} parameter in the bivariate gamma distribution. Note that, depending on the model type, might not be necessary to provide it.}

\item{maxit}{A parameter that controls the number of maximum iteration in the EM algorithm. The default is 100.}

\item{tol}{A parameter that controls the convergence tolerance in the EM algorithm. The default is 1e-5.}

\item{verbose}{A logical controlling whether estimations in each EM iteration are shown in the fitting process. The default is TRUE.}
}
\value{
An object of class \code{BGR} providing the estimation results.
  The details of the output components are:
  \item{call}{The matched call.}
  \item{coefficients}{The estimated coefficients.}
  \item{alpha1}{The estimated alpha1 values.}
  \item{alpha2}{The estimated alpha2 values.}
  \item{alpha3}{The estimated alpha3 values.}
  \item{beta}{The estimated beta values.}
  \item{fitted.values}{The fitted values of the regression.}
  \item{loglike}{The final estimated maximum log-likelihood value.}
  \item{ll}{The sequence of log-likelihood values in the EM algorithm fitting process.}
  \item{df}{Number of estimated parameters.}
  \item{AIC}{AIC values.}
  \item{BIC}{BIC values.}
  \item{iter}{Total iteration numbers.}
  \item{formula}{The formulas used in the regression.}
  \item{y}{The input response data.}
  \item{n}{The number of observations in the data.}
  \item{Model.Matrix}{The used model matrix for each regression formula.}
  \item{trace}{All estimated coefficients and alpha, beta values in the EM algorithm.}
}
\description{
Bivariate gamma regression models for three model types: EE, EI, IE. Models are estimated by EM algorithm.
}
\examples{

\donttest{
mod1 <- BGR(modelName = "EE",
            y = c("y1","y2"), data = fullsim,
            f1 = ~ w1 + w2,
            f2 = ~ w2 + w3,
            f3 = ~ w1 + w2 + w3,
            f4 = ~ w1 + w2 + w3,
            verbose= FALSE)
mod1

mod2 <- BGR(modelName = "EI",
            y = c("y1","y2"), data = fullsim,
            f1     = ~ w1 + w2,
            f2     = ~ w2 + w3,
            f3     = ~ w1 + w2 + w3,
            verbose= FALSE)
mod2
mod3 <- BGR(modelName = "IE",
            y = c("y1","y2"), data = fullsim,
            f4     = ~ w1 + w2 + w3,
            verbose= FALSE)
mod3
}

}
