% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbivgamma.R
\name{dbivgamma}
\alias{dbivgamma}
\title{Density of bivariate gamma distribution given parameters values}
\usage{
dbivgamma(y, alpha, beta, log = FALSE)
}
\arguments{
\item{y}{vector of quantiles}

\item{alpha}{vector of alpha values in bivariate gamma distribution}

\item{beta}{value of beta in bivariate gamma distribution}

\item{log}{logical; if TRUE, probabilities p are given as log(p). Default is FALSE.}
}
\description{
Density of bivariate gamma distribution given parameters values
}
