% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBGC.R
\name{MBGC}
\alias{MBGC}
\title{Mixture of bivariate gamma distributions clustering}
\usage{
MBGC(data, G, maxit, tol, start = "random.start",
  starting.value = NULL, Aitken = FALSE, verbose = TRUE)
}
\description{
Estimation using EM algorithm for mixture of bivariate gamma distributions
}
