% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.MBGR.R
\name{predict.MBGR}
\alias{predict.MBGR}
\title{Predict Method for Bivariate Gamma Regressions}
\usage{
\method{predict}{MBGR}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted model object of class \code{"BGR"} or \code{"MBGR"}}

\item{newdata}{a data frame that prediction is based upon}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list with components (depends on model class):
  \item{fit}{predictions.}
  \item{alpha1.fit}{predicted alpha1 values.}
  \item{alpha2.fit}{predicted alpha2 values.}
  \item{alpha3.fit}{predicted alpha3 values.}
  \item{beta.fit}{predicted beta values.}
  \item{tau.fit}{predicted mixing proportion values, if covariates enter the gating network.}
}
\description{
Obtains predictions from the fitted bivariate gamma regression models.
}
\examples{

\donttest{
m1 <- MBGR(modelName = "VC", G=2,
           y=c("y1","y2"), data = fullsim,
           f1     = ~ w1 + w2,
           f2     = ~ w2 + w3,
           f3     = ~ w1 + w2 + w3,
           f4     = ~ w1 + w2 + w3,
           gating = "C", verbose=FALSE)
fitted1 <- predict(m1, newdata=fullsim)
}

}
